Attempt to clean up some of the ChefSpec boilerplate
