--- spec/support/cookbooks/resource_dnsmasq_local_config_test/metadata.rb
@@ -1,11 +1,0 @@
-# Encoding: UTF-8
-
-name 'resource_dnsmasq_local_config_test'
-maintainer 'Jonathan Hartman'
-maintainer_email 'jonathan.hartman@socrata.com'
-license 'apache2'
-description 'Resource test wrapper'
-long_description 'Resource test wrapper'
-version '0.0.1'
-
-depends 'dnsmasq-local'
